Remove randudant code, renaming
