+++ react-basic-practice/src/components/CustomerForm/index.tsx
