+++ react-basic-practice/src/constants/modal.ts
@@ -2,4 +2,5 @@ export const MODAL_TITLES = {
   ADD_CUSTOMER: "Add Customer",
   EDIT_CUSTOMER: "Edit Customer",
   VIEW_CUSTOMER: "View Customer",
+  DELETE_CUSTOMER: "Delete Customer",
 };
